+++ Source/swiftlint/Commands/Analyze.swift
@@ -24,6 +24,7 @@ extension SwiftLint {
                 paths: allPaths,
                 useSTDIN: false,
                 configurationFiles: common.config,
+                parentConfig: common.parentConfig,
                 strict: common.leniency == .strict,
                 lenient: common.leniency == .lenient,
                 forceExclude: common.forceExclude,