+++ Source/swiftlint/Commands/Lint.swift
@@ -41,6 +41,7 @@ extension SwiftLint {
                 paths: allPaths,
                 useSTDIN: useSTDIN,
                 configurationFiles: common.config,
+                parentConfig: common.parentConfig,
                 strict: common.leniency == .strict,
                 lenient: common.leniency == .lenient,
                 forceExclude: common.forceExclude,